Adds ward/inpatient area flags to locations
Adds `isWardOrFlowArea` and `isCoreInpatientArea` flags to the Department and DepartmentState entities.

These flags indicate whether a department is a ward or flow area and whether it is a core inpatient area, respectively.

Adds logic to update these flags in the database when processing location metadata.
Also adds validation to ensure that previous values of these flags match the database values before invalidating previous states.